+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/module/death/respawn/client/PlayerRendererMixin_GhostlyEffect.java
@@ -4,7 +4,7 @@ import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.llamalad7.mixinextras.sugar.Local;
 import insane96mcp.insanelib.core.feature.Feature;
-import insane96mcp.insanesurvivaloverhaul.module.death.respawn.Ghostly;
+import insane96mcp.insanesurvivaloverhaul.module.respawn.Ghostly;
 import net.minecraft.client.model.geom.ModelPart;
 import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.renderer.RenderType;