+++ src/main/java/insane96mcp/insanesurvivaloverhaul/module/respawn/echopillar/EchoPillar.java
@@ -1,4 +1,4 @@
-package insane96mcp.insanesurvivaloverhaul.module.death.respawn;
+package insane96mcp.insanesurvivaloverhaul.module.respawn.echopillar;
 
 import insane96mcp.insanelib.core.JsonFeature;
 import insane96mcp.insanelib.core.feature.LoadFeature;
@@ -12,6 +12,7 @@ import insane96mcp.insanesurvivaloverhaul.data.generator.ISOItemTagsProvider;
 import insane96mcp.insanesurvivaloverhaul.event.DeathPenaltyEvent;
 import insane96mcp.insanesurvivaloverhaul.module.ISOModules;
 import insane96mcp.insanesurvivaloverhaul.module.misc.Packs;
+import insane96mcp.insanesurvivaloverhaul.module.respawn.Ghostly;
 import insane96mcp.insanesurvivaloverhaul.setup.ISORegistries;
 import insane96mcp.insanesurvivaloverhaul.setup.SimpleBlockWithItem;
 import net.minecraft.ChatFormatting;
@@ -38,7 +39,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
 
-@LoadFeature(module = ISOModules.DEATH, description = "Adds Echo Pillars: ancient structures that act as spawn anchors and grant effects on respawn.", enabledByDefault = false)
+@LoadFeature(module = ISOModules.RESPAWN, description = "Adds Echo Pillars: ancient structures that act as spawn anchors and grant effects on respawn.", enabledByDefault = false)
 public class EchoPillar extends JsonFeature {
 	public static final TagKey<Block> ECHO_PILLAR_BLOCKS_TO_ROT = ISOBlockTagsProvider.create("structures/echo_pillar/blocks_to_rot");
 	public static final TagKey<Item> ECHO_PILLAR_CATALYST = ISOItemTagsProvider.create("echo_pillar_catalyst");